pyproject.toml
[tool.poetry]
name = "langchain-googledrive"
version = "0.0.0"
description = "This is a temporary project while I wait for my langchain [pull-request](https://github.com/hwchase17/langchain/pull/5135) to be validated."
authors = ["Philippe PRADOS"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://www.github.com/pprados/langchain-googledrive"
packages = [{ include = "langchain_googledrive" }]

[tool.poetry.dependencies]
python = "^3.8.1"
langchain_core = ">=0.2.1"
langchain_community = ">=0.2.1"
google-api-python-client = "^2.97"
google-auth-httplib2 = ">=0.1"
google-auth-oauthlib = ">=1.0"

[tool.poetry.extras]
all = [
    "pdf2image",
    "pypandoc_binary",
    "torch",
#    "detectron2",
    "pytesseract",
    "unstructured"
]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
playwright = "^1.28.0"
setuptools = "^68.1.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
# The only dependencies that should be added are
# dependencies used for running tests (e.g., pytest, freezegun, response).
# Any dependencies that do not meet that criteria will be removed.
pytest = "^7.3.0"
pytest-cov = "^4.0.0"
pytest-dotenv = "^0.5.2"
pytest-watcher = "^0.2.6"
pytest-asyncio = "^0.20.3"
pytest-mock = "^3.10.0"
pytest-socket = "^0.6.0"
twine = "^4.0.0"
langchain-openai = ">=0.1.7"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
mypy = "^0.991"
ruff = "^0.0.249"
types-toml = "^0.10.8.1"
types-redis = "^4.3.21.6"
types-pytz = "^2023.3.0.0"
black = "^23.1.0"
types-chardet = "^5.0.4.6"
mypy-protobuf = "^3.0.0"
types-google-cloud-ndb = "^2.0"
types-docutils = "^0.20"
types-pyyaml = "^6.0.12.2"
types-requests = "^2.28.11.5"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
autodoc_pydantic = "^1.8.0"
myst_parser = "^0.18.1"
nbsphinx = "^0.8.9"
sphinx = "^4.5.0"
sphinx-autobuild = "^2021.3.14"
sphinx_book_theme = "^0.3.3"
sphinx_rtd_theme = "^1.0.0"
sphinx-typlog-theme = "^0.8.0"
sphinx-panels = "^0.6.0"
toml = "^0.10.2"
myst-nb = "^0.17.1"
linkchecker = "^10.2.1"
sphinx-copybutton = "^0.5.1"
nbdoc = "^0.0.82"

[tool.poetry.group.codespell]
optional = true

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.5"

[tool.codespell]
skip = '.git,*.pdf,*.svg,*.pdf,*.yaml,*.ipynb,poetry.lock,*.min.js,*.css,package-lock.json,example_data,_dist,examples'
# Ignore latin etc
ignore-regex = '.*(Stati Uniti|Tense=Pres).*'
# whats is a typo but used frequently in queries so kept as is
# aapply - async apply
# unsecure - typo but part of API, decided to not bother for now
ignore-words-list = 'momento,collison,ned,foor,reworkd,parth,whats,aapply,mysogyny,unsecure,damon,crate'


[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
